+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Added webflow to image CSP
 - Will now log the extension version to the console on startup
 - The get help button now shows the top committer name instead of "on Slack"
+- Fixed an error where the command would break even when results were fetched
 
 ## [1.1.3]
 - Fixed some instances of the "Get help on Slack" running search